jws: take the import copy-helper length as size_t
cjose_jws_import() computes the header/payload/signature segment offsets as
size_t (commit 2fef0f8, to avoid truncating them for a >2GiB serialization),
but then passed each length into _cjose_jws_strcpy() through an int
parameter, truncating them straight back. For an oversized cser where a
segment length's low 32 bits are small, the buffer is allocated/copied at the
truncated size while the following cjose_base64url_decode() is handed the full
size_t length, over-reading past the allocation.

Change the parameter to size_t, completing the earlier size_t hardening. All
callers already pass size_t lengths bounded by cser_len.

+++ src/jws.c
@@ -756,7 +756,7 @@ bool cjose_jws_export(cjose_jws_t *jws, const char **compact, cjose_err *err)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-static bool _cjose_jws_strcpy(char **dst, const char *src, int len, cjose_err *err)
+static bool _cjose_jws_strcpy(char **dst, const char *src, size_t len, cjose_err *err)
 {
     *dst = (char *)cjose_get_alloc()(len + 1);
     if (NULL == *dst)
